kraken-bot/requirements.txt
flask
krakenex
requests
pandas
numpy
nltk
libsass
python-dotenv
ratelimit